+++ tests/asyncio/test_connection.py
@@ -186,3 +186,22 @@ async def test_run_callable_warning(asyncio_engine):
     async with asyncio_engine.connect() as conn:
         await conn._worker.run(thread_fn, conn)
         assert thread_called
+
+
+@pytest.mark.asyncio
+async def test_run_visitor_exception(asyncio_engine, mytable):
+    thread_called = False
+
+    def thread_fn(conn):
+        nonlocal thread_called
+
+        with pytest.raises(AttributeError, match='Did you try to use'):
+            mytable.create(conn)
+
+        mytable.create(conn.sync_connection)
+
+        thread_called = True
+
+    async with asyncio_engine.connect() as conn:
+        await conn._worker.run(thread_fn, conn)
+        assert thread_called